requirements-dev.txt
-r requirements.txt

# Linting and formatting
yapf
pylint
toml

# Documentation
numpydoc
Sphinx==5.3.0
sphinx-autoapi
sphinx-copybutton
sphinx-gallery
sphinx-rtd-theme
esbonio
